Move i18n files into a new directory
