+++ src/packages/runtime-react-support/PackageContext.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023-2025 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
 import { createContext } from "react";
-import type { PackageIntl } from "@open-pioneer/runtime/i18n";
+import type { PackageIntl } from "@open-pioneer/runtime";
 import type { Service } from "@open-pioneer/runtime/Service";
 import type { UseServiceOptions } from "@open-pioneer/runtime/react-integration/hooks";
