+++ src/packages/runtime/service-layer/verifyDependencies.test.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023-2025 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
 import { assert, expect, it } from "vitest";
-import { createEmptyI18n } from "../i18n";
+import { createEmptyPackageIntl } from "../i18n";
 import { expectError } from "../test-utils/expectError";
 import { InterfaceSpec, ReferenceSpec } from "./InterfaceSpec";
 import { ReadonlyServiceLookup } from "./ServiceLookup";
@@ -406,7 +406,7 @@ function mockServices(data: ServiceData[]): ServiceRepr[] {
             name,
             packageName,
             factory: createConstructorFactory(clazz),
-            intl: createEmptyI18n(),
+            intl: createEmptyPackageIntl(),
             dependencies,
             interfaces
         });
